+++ apps/question-bank/src/main/java/gov/cdc/nbs/questionbank/valueset/concept/ConceptCreator.java
@@ -1,4 +1,4 @@
-package gov.cdc.nbs.questionbank.valueset;
+package gov.cdc.nbs.questionbank.valueset.concept;
 
 import java.time.Instant;
 import java.util.Optional;
