+++ apps/question-bank/src/test/java/gov/cdc/nbs/questionbank/question/QuestionFinderTest.java
@@ -13,7 +13,7 @@ import java.util.*;
 import gov.cdc.nbs.questionbank.question.exception.UniqueQuestionException;
 import gov.cdc.nbs.questionbank.question.request.QuestionValidationRequest;
 import gov.cdc.nbs.questionbank.question.request.QuestionValidationRequest.Field;
-import gov.cdc.nbs.questionbank.valueset.ConceptFinder;
+import gov.cdc.nbs.questionbank.valueset.concept.ConceptFinder;
 import gov.cdc.nbs.questionbank.valueset.model.Concept;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
